requirements.txt
bio==1.5.9
biopython==1.81
keras==2.12.0
numpy==1.23.5
pandas==2.0.0
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
sklearn==0.0.post1
sparse==0.14.0
tensorflow==2.12.0
